Go-Audit - Refactoring to allow the components of go-audit to be used as libraries Ticket #: EM-1631
Root Cause Analysis: Go-audit source provides an executable to be run as-is on a given system. We
need to modify the main program substantially so it makes sense to write our own.  Unfortunately the
go-audit source is not packaged to allow the various components to be included as libraries.

Fix/Implementation: Refactor the go-audit source code to allow the individual components to used separately
and add a logging abstraction layer.

Reviewed By: Matt Madison